Refactor APR API into multiple files and implemented zod (#286)
* Refactor APR API into multiple files and implemented zod
